default-signal-handler takes 5 args: bus path svc iface mber
However maybe it's a bad idea because a dbus_message has more stuff, so
the handler doesn't get enough info to actually handle the signal.
But what kind of interface is that even if it could?  The idea is this
egg is to be a nice high-level abstraction, not a way of digging around
in the bowels of how libdbus does things.  And if we pass the whole message 
to this handler, it's a C object, so what if the handler tries to keep it
for too long.  Have never addressed GC yet anyway.